Change steam avatar url from https to http
The scheme part for links is hardcoded in smf and with a non https site, Load.php looks for avatar strings that start with http not https, which is what steam returns during openid login. When that fails, it assumes the https url is a file in the local avatars folder.

Steam avatar urls can be specified either http or https so changing them to http is easier than changing the avatar loading logic in Load.php.